+++ src/pages/homepage/index.js
@@ -1,0 +1,13 @@
+import { Box } from "@chakra-ui/react";
+import React from "react";
+import Header from "../../components/header";
+
+const Homepage = () => {
+  return (
+    <Box>
+      <Header />
+    </Box>
+  );
+};
+
+export default Homepage;
